Added Util function
Added util and moved some function in.
Also changed predict to load the model
